+++ index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="js/main.js"></script>
+
   <title>Кекстаграм</title>
 </head>
 
@@ -165,6 +165,6 @@
       </div>
     </section>
   </template>
-
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
